refactor: move ValidatedModel into separate file
