+++ apps/rubrics/models/rubric.py
@@ -5,6 +5,8 @@ from django.db import models, transaction
 from django.db.models import Max, Q
 from django.utils import timezone
 
+from .base import ValidatedModel
+
 
 def value_is_numeric(value) -> bool:
     try:
@@ -13,39 +15,6 @@ def value_is_numeric(value) -> bool:
         return False
     else:
         return True
-
-
-class ValidatedModel(models.Model):
-    """Adds additional validation to a model on save.
-
-    This class defines a framework for adding additional validation
-    on save of an object. The validation is written in such a way that it
-    can easily be re-used by forms. As such, the validation functions expect
-    dictionaries.
-    """
-
-    class Meta:
-        abstract = True
-
-    def save(self, **kwargs):
-        data = self.get_field_dict()
-        self.validate(**data)
-        self.validate_instance(**data)
-
-        super(ValidatedModel, self).save(**kwargs)
-
-    def get_field_dict(self) -> dict:
-        data = {}
-        for field in self._meta.fields:
-            data[field.name] = getattr(self, field.name)
-        return data
-
-    def validate_instance(self, **fields):
-        pass
-
-    @classmethod
-    def validate(cls, **fields):
-        pass
 
 
 class Rubric(ValidatedModel):
